cleanup: mixer
