+++ src/base/mixing.h
@@ -7,14 +7,6 @@
  * outbuf must big enough to hold output_channels*samples_to_do */
 void mix_vgmstream(sample_t *outbuf, int32_t sample_count, VGMSTREAM* vgmstream);
 
-/* internal mixing pre-setup for vgmstream (doesn't imply usage).
- * If init somehow fails next calls are ignored. */
-void* mixer_init(int channels);
-void mixer_free(void* mixer);
-void mixer_update_channel(void* mixer);
-void mixer_process(void* _mixer, sample_t *outbuf, int32_t sample_count, int32_t current_pos);
-bool mixer_is_active(void* mixer);
-
 /* Call to let vgmstream apply mixing, which must handle input/output_channels.
  * Once mixing is active any new mixes are ignored (to avoid the possibility
  * of down/upmixing without querying input/output_channels). */
@@ -41,4 +33,4 @@ void mixing_macro_crosslayer(VGMSTREAM* vgmstream, int max, char mode);
 void mixing_macro_downmix(VGMSTREAM* vgmstream, int max /*, mapping_t output_mapping*/);
 
 
-#endif /* _MIXING_H_ */
+#endif
